Create Standing Stadium2.cpp
This program prompts the user to enter a name of a movie and the number of tickets they would like to buy. Then it will show the box office report of the profit of both the Gross Box Office and Net Box Office, and the amount paid to the Distributor.