Create Apartments.java
Create a class named Apartment that holds an apartment number, number of bedrooms, number of baths, and rent amount. Create a constructor that accepts values for each data field. Also, create a get method for each field. Write an application that creates at least five Apartment objects(apartment number, number of bedrooms, number of baths, and monthly rent amount). Then prompt the user to enter a minimum number of bedrooms required, a minimum number of baths required, and a maximum rent the user is willing to pay. Display data for all the Apartment objects that meet the user's criteria or an appropriate message if no such apartments are available.